Rewrite group-deployments to sort PRs by production release
Grouping PRs by comparing merge times with deployment times doesn't
really work since it ignores the fact that master branch is going
to production through staging: so if staging is released to production
after a delay PRs merged during that delay shouldn't be considered part
of the release.

New approach is to sort PRs by merge time and then split the sequence
each time we find a commit that was released to production. This still
assumes that all PRs are merged to the same branch and that branch is
the one being deployed. I could add branch filters in the future, but
I'm not sure what else might be affected by this.

The return value is a sequence where each element is a sequence of
PRs that were released to production at the same time. This means
we can concatenate this with similar sequences from other repos so
that:

* we can sort PRs from multiple repos in the same list by the time
  they were merged
* PRs that were released to production at the same time are grouped
  together (so that it's easy to see what was shipped in each release)